Additional testing code.

+++ src/test/java/de/carne/certmgr/test/certs/io/CertReadersWritersTest.java
@@ -19,12 +19,12 @@ package de.carne.certmgr.test.certs.io;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
+import java.io.Writer;
 import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.security.Security;
 import java.util.Arrays;
-import java.util.List;
 
 import org.bouncycastle.jce.provider.BouncyCastleProvider;
 import org.junit.Assert;
@@ -38,7 +38,6 @@ import de.carne.certmgr.certs.io.IOResource;
 import de.carne.certmgr.certs.spi.CertReader;
 import de.carne.certmgr.certs.spi.CertWriter;
 import de.carne.certmgr.test.Tests;
-import de.carne.util.Exceptions;
 
 /**
  * Test Certificate Readers and Writers.
@@ -60,22 +59,14 @@ public class CertReadersWritersTest {
 	 */
 	@Test
 	public void testReadersAndWriters() throws IOException {
-		// List<String> testResources = Arrays.asList("test.crt", "test.key",
-		// "test.csr", "test.crl", "test.pem");
-		List<String> testResources = Arrays.asList("test.crt", "test.pem");
+		for (CertReader reader : CertReaders.REGISTERED.providers()) {
+			URL testResourceURL = getClass().getResource(reader.providerName() + ".dat");
 
-		for (String testResource : testResources) {
-			URL testResourceURL = getClass().getResource(testResource);
-			CertObjectStore certObjects = CertReaders.readURL(testResourceURL, Tests.password());
-
-			Assert.assertNotNull(certObjects);
-			Assert.assertTrue(certObjects.size() > 0);
-
-			for (CertWriter writer : CertWriters.REGISTERED.providers()) {
+			if (testResourceURL != null) {
 				Path testPath = Files.createTempFile(getClass().getSimpleName(), null);
 
 				try {
-					testReaderAndWriter(writer, testPath, certObjects);
+					testReaderAndWriter(reader, testResourceURL, testPath);
 				} finally {
 					Files.delete(testPath);
 				}
@@ -83,38 +74,66 @@ public class CertReadersWritersTest {
 		}
 	}
 
-	private void testReaderAndWriter(CertWriter writer, Path testPath, CertObjectStore certObjects) throws IOException {
-		System.out.println("Testing provider: " + writer.providerName());
+	private void testReaderAndWriter(CertReader reader, URL testResourceURL, Path testPath) throws IOException {
+		System.out.println("Testing I/O provider: " + reader.providerName());
 
-		CertReader reader = CertReaders.REGISTERED.get(writer.providerName());
+		CertWriter writer = CertWriters.REGISTERED.get(reader.providerName());
 
-		System.out.println(reader != null ? "(Reading and Writing)" : "(Writing only)");
-		System.out.println(writer.fileType());
-		System.out.println(Arrays.toString(writer.fileExtensionPatterns()));
-		System.out.println("isCharWriter: " + writer.isCharWriter());
-		System.out.println("isEncryptionRequired: " + writer.isEncryptionRequired());
-
-		boolean certObjectsWritten = false;
-
-		try (IOResource<OutputStream> out = IOResource.newOutputStream(testPath.toString(), testPath)) {
-			writer.writeBinary(out, certObjects);
-		} catch (UnsupportedOperationException e) {
-			Exceptions.ignore(e);
-			Assert.assertTrue(writer.isEncryptionRequired());
-		}
-		try (IOResource<OutputStream> out = IOResource.newOutputStream(testPath.toString(), testPath)) {
-			writer.writeEncryptedBinary(out, certObjects, Tests.password());
-			certObjectsWritten = true;
+		System.out.println(reader.fileType());
+		System.out.println(Arrays.toString(reader.fileExtensionPatterns()));
+		if (writer != null) {
+			System.out.println("isCharWriter: " + writer.isCharWriter());
+			System.out.println("isEncryptionRequired: " + writer.isEncryptionRequired());
 		}
 
-		if (reader != null && certObjectsWritten) {
-			try (IOResource<InputStream> in = IOResource.newInputStream(testPath.toString(), testPath)) {
-				CertObjectStore readCertObjects = reader.readBinary(in, Tests.password());
+		CertObjectStore readCertObjects1 = CertReaders.readURL(testResourceURL, Tests.password());
 
-				// Assert.assertNotNull(readCertObjects);
-				// Assert.assertEquals(certObjects.size(),
-				// readCertObjects.size());
+		Assert.assertNotNull(readCertObjects1);
+
+		CertObjectStore readCertObjects2;
+
+		try (IOResource<InputStream> in = new IOResource<>(testResourceURL.openStream(), reader.providerName())) {
+			readCertObjects2 = reader.readBinary(in, Tests.password());
+
+			Assert.assertNotNull(readCertObjects2);
+			Assert.assertEquals(readCertObjects1.size(), readCertObjects2.size());
+		}
+		if (writer != null) {
+			if (!writer.isEncryptionRequired()) {
+				try (IOResource<OutputStream> out = IOResource.newOutputStream(writer.providerName(), testPath)) {
+					writer.writeBinary(out, readCertObjects2);
+				}
+				verifyWriterOutput(readCertObjects2, reader, testPath);
+				if (writer.isCharWriter()) {
+					try (IOResource<Writer> out = new IOResource<>(Files.newBufferedWriter(testPath),
+							writer.providerName())) {
+						writer.writeString(out, readCertObjects2);
+					}
+					verifyWriterOutput(readCertObjects2, reader, testPath);
+				}
 			}
+			try (IOResource<OutputStream> out = IOResource.newOutputStream(writer.providerName(), testPath)) {
+				writer.writeEncryptedBinary(out, readCertObjects2, Tests.password());
+			}
+			if (writer.isCharWriter()) {
+				try (IOResource<Writer> out = new IOResource<>(Files.newBufferedWriter(testPath),
+						writer.providerName())) {
+					writer.writeEncryptedString(out, readCertObjects2, Tests.password());
+				}
+				verifyWriterOutput(readCertObjects2, reader, testPath);
+			}
+			verifyWriterOutput(readCertObjects2, reader, testPath);
+		}
+		System.out.println();
+	}
+
+	private static void verifyWriterOutput(CertObjectStore readCertObjects, CertReader reader, Path testPath)
+			throws IOException {
+		try (IOResource<InputStream> in = IOResource.newInputStream(testPath.toString(), testPath)) {
+			CertObjectStore readCertObjects2 = reader.readBinary(in, Tests.password());
+
+			Assert.assertNotNull(readCertObjects2);
+			Assert.assertEquals(readCertObjects.size(), readCertObjects2.size());
 		}
 	}
